Fixed print issue
There was an issue where content with dashes (-) would act like the dashes were <br> and skip the next part to the next line. This is fixed.

Also made it so when the print window is closed the content is cleared and you're left with a blank A4 area to start over again.